requirements.txt
Django==1.11.2
django-ratelimit==1.0.1
django-simple-search==1.0.2
django-widget-tweaks==1.4.1
html5lib==1.0b10
httplib2==0.10.3
olefile==0.44
Pillow==4.1.1
PyPDF2==1.26.0
pytz==2017.2
reportlab==3.4.0
six==1.10.0
webencodings==0.5.1
xhtml2pdf==0.2b1
xlrd==1.0.0
